Export and Import OPML workers send emails immediately, instead of enqueuing a job to send them.
Interactivity is not critical in these cases because emails are being sent from a background job, so this should not degrade user experience.

This should fix a problem when exporting feeds: for some users the export failed when trying to call deliver_later to send an email success; the root cause is that the email or its attachment had characters that could not be serialized into JSON when enqueuing the job. The simplest fix is just to deliver the email synchronously, so that it doesn't get serialized to JSON.