In ActiveAdmin, in the subscribe jobs table (part of the show user view), display a column with the fetch_url instead of the feed id.
It is interesting to see the fetch_url entered by the user for each job, to detect possible issues. On the other hand the feed id is not very interesting (it will only have a value if the job finished successfully), and in any case there is a column with a link to show the feed (again, only if the job finished successfully).