Restored old html sanitizer behavior by including "rails-deprecated_sanitizer" gem.
The new rails sanitizer behaves badly in some cases: for extremely complicated (but actually occurring) entries, it raises a "stack level too deep" error. It is not currently suitable for my needs; I suspect the problem is actually in the underlying Loofah gem, so I don't think it will be trustworthy in the future either.

Using the rails-deprecated_sanitizer gem is a temporary measure, because it will stop being maintained in Rails 5. I need to find a supported and well-behaving sanitization gem in the near future.