Fetch rack-cache gem from my own github fork.
The current rack-cache stable (and github master head) has a bug: if the response from a server has an Expires header which does not conform to RFC2616 (e.g. because the day has one digit, when the RFC mandates two), an error is raised (ArgumentError, raised by Time.httpdate when passed an unparseable date) which makes the whole request fail.

There is a long-standing PR which fixes the problem, making rack-cache ignore the unparseable header instead of raising an error. However the PR has not been merged for six months, there is no feedback from repo owner, and generally it looks like the project is dormant or abandandoned.

I've forked the repo and applied the commit from the PR. I don't expect any further development in the gem, so it shouldn't be a burden to maintain this fork.

The PR which sadly has not been merged is:

  https://github.com/rtomayko/rack-cache/pull/97/commits

I reported the problem in the rack-cache bugtracker here:

  https://github.com/rtomayko/rack-cache/issues/107