Fix bug: when reading all entries in a folder just after unsubscribing from a feed in the folder, sometimes an error appeard.
This was caused by a timing problem. Actual unsubscribing is done in a background job. If the user was quick enough, EntriesController#index could be invoked before Sidekiq had time to perform the unsubscribing. In this case it was possible that entries from the (not yet) unsubscribed feed were present in the list returned by ActiveRecord; but by the time the view was rendered, when Entry#read_by? was invoked for those entries, the feed could already have been unsubscribed (and all read/unread state information for its entries and for the user would already be deleted). In this case an error would be raised during view rendering (because .read method would be invoked on a nil EntryState instance), which would mean an HTTP 500 would be returned and an error shown to the user.

The fix is when rendering the entries_index view, to invoke Entry#read_by? as soon as possible for each entry passed to the view; and then to rescue NotSubscribedError in the view, inside the entries loop. This way any entry which has no EntryState information will just be skipped (not present in the returned JSON) and the error will be silently swallowed.

This means that some entries passed to the index view will be ignored, as they should be; this requires complicating a bit the logic that outputs "," caracters between entries in the JSON view, because it is no longer guaranteed that the first and last entries passed in the view will be present as first and last entries in the rendered JSON.