Fix for images inside tables being displayed at native width, instead of scaled responsively.
There is an inconsistence between browsers in how max-width for images inside tables is handled. Chrome handles this perfectly but firefox and opera display the image at its original width, which means that large images are cropped (part of the image is outside the viewport).

Reported here among others:

  https://bugzilla.mozilla.org/show_bug.cgi?id=975632
  http://stackoverflow.com/questions/18846744/responsive-images-in-tables-bootstrap-3/1885903
  https://bugzilla.mozilla.org/show_bug.cgi?id=434230

The suggested fix is to set the table layout to fixed and to set a table width of 100% (setting a max-width for the table instead doesn't work). This means that tables in entries now fill the whole width of the entries column.

This fixes the original problem (images inside tables are now responsively scaled) but it may introduce problems in cases where tables are not expected to have 100% width. I will have to keep an eye out for such cases, but for now I haven't found any.