WIP: Replace last_modified with etag headers when requesting the state of subscribe and refresh jobs.
The last-modified http header offers a precision of 1 second which is not enough in some cases, causing the user to see a wrong state for his jobs at least for a time. The fix is to use an etag with greater precission. The current time is calculated in seconds with floating point precision (fractions of a second with a lot of precision), an MD5 hash of that number is obtained, and that value is stored in the User model to be used as an Etag.