Fix regression: when saving or destroying a RefreshFeedJobState instance, check that the user exists before updating his refresh_feed_jobs_etag attribute.
This is specially important when destroying the instance; this may have been triggered by ActiveRecord callbacks after destroying the user, and this would mean that the user no longer exists. If we don't validate that the user exists before updating the etag, we'd end up invoking methods on a nil variable.