Fix tests to adapt to recent changes in lazy-loading images.
It is no longer guaranteed that after attempting to lazy load an image its URL scheme will remain unchanged. The new lazy loading code changes the URL from http to https and vice versa for a second loading attempt if the first attempt to load the image fails. Tests were not designed to support this.