Check every second if a minute has passed since the last feeds refresh. If so, refresh feeds.
This replaces the old behavior in which simply a feed refresh is triggered every minute.

The downside of the old behavior was that sometimes, specially in mobile browsers, the browser loaded the old state of the app which could be outdated (wrong entries counts etc) because the app had been closed hours ago; in this case the user could see the wrong entries during the first minute. Worse yet, if he performed actions that reset the refresh timer (like simply opening an entry), he could see outdated feeds information for longer than a minute.

The new behavior avoids this by checking every second if more than one minute has passed, making a refresh necessary. This way the user will see outdated information for at most one second before an automatic refresh of feeds is triggered.

For this to work correctly, the function that resets the refresh timer after certain actions (unsubscribe from a feed, change entries state etc) has been modified so that the timer is not reset if more than 90 seconds have passed since the last refresh. This means that if more than 90 seconds have passed since the last refres, feeds will be refreshed in the next second regardless what actions the user performs.