Fix a bug that could cause the client to enter a loop of feed refreshes every second.
When a feed refresh is requested from the server, the timestamp of the last refresh is set to the current time, unconditionally. This is done directly instead of passing through the reset_timer() function; the reason is that the function does not modify the timestamp if its more than 90 seconds in the past. This is intended so that refreshes are not delayed when the user e.g. opens entries if the last refresh was too far in the past; but this behavior is not desirable when a feed refresh actually happens, in this case the timestamp must be updated unconditionally.