Folders are also refreshed by the client every minute, not just feeds.
This reduces the risk of folders being out of synch if the user uses several clients (mobile and desktop browser, e.g.). This is specially important if the user has installed the app as fullscreen in his smartphone, and therefore has no easy way to refresh the page forcing a refresh of folders.