Store the timestamp of the last change in subscriptions for each user, instead of the md5 hash of the timestamp.
This allows greater precision in http caching. The last_modified header is simply the subscriptions_updated_at attribute (converted to the format allowed by the http spec, with a 1-second precision) and the etag is calculated on the fly as the md5 hash of the timestamp (with a precision of several decimals of a secnd).