Explicitly set size of the Redis connections pool used in Sidekiq, both client and server.
The server needs (concurrency+2) connections; currently concurrency is 3 threads, so the size is set to 5.

The client needs just 1 connection (by default Rails gives a size of 5, which is overkill). This is per-process; currently Puma uses 2 workers (=2 processes), which means a total of 2 connections from the web server.

The biggest win here is limiting the size of the client pool, which was too big with the default settings. This may help resolving recent issues with sidekiq not being able to get a redis connection.

For more about this see:

  https://github.com/mperham/sidekiq/issues/117