When fetching a feed, strip each entry guid before checking if it already exists in the database.
This is necessary because the Entry model strips the guid before saving (actually before validation, but whatever). So if e.g. an entry with guid "aa " is received, it will be saved with guid "aa". When next fetched we have to strip the guid before searching in the db; otherwise if we search for an entry with guid "aa " it won't be found, the program will think that the entry is not yet saved and will attempt to save it. The result will be a "guid already taken" error being raised.

The above described situation is not a big problem, because the end result is that the already existing entry won't be saved again, which is the desired behavior. However the logs will be cluttered with this error over and over again (the raised error becomes a log with high priority being printed). Better to fix the logic that checks if the entry already exists; if it does, a log is printed with "debug" priority, which is usually below the threshold for production. This means that in production already existing entries will be silently ignored, instead of cluttering the logs with useless errors.