When saving a new entry, do not print the backtrace if an ActiveRecord validation error is raised.
An AR validation error in this point means an invalid entry (missing URL, etc). Writing the backtrace to the log is just clutter, the only useful information is the error message, which tells us which validations have failed exactly.

In the case other errors are raised the backtrace is written to the log, to help debugging unexpected errors.