Added boolean attribute named "free" to User model.
It is mandatory and defaults to false.

Most users will have this attribute set to false. This means that they are regular users who at some point are required to pay to access the app (even during any unpaid trial period the attribute is set to false, the unpaid trial will be controlled with other attributes).

For those fortunate users who have the attribute set to true, they will be able to use the app without having to pay, for as long as the attribute is set to true.

Only admins can change this attribute value through the admin interface. It is intended for people who through special circumstances (e.g. having participated in the closed beta) are granted free access, probably forever.