Support for a URL blacklist in a config file. Users won't be able to subscribe to feeds in the blacklist.
Implemented: The blacklist is loaded from a YAML config file on startup. Before feed validation (i.e. before saving a feed), the feed's url and fetch_url are compared to the blacklist. If either is found to be a sub-url of any of the URLs in the blacklist, a BlacklistedUrlError is raised, preventing the feed from being saved.

Not yet implemented: notify the user of the reason why he cannot subscribe, instead of displaying a generic error.