WIP: when a user requests a subscribe job to be enqueued, the url is checked against the blacklist first.
If the url matches an element of the blacklist no job will be enqueued and the job state will be set to ERROR.

Not yet implemented: raise an error and return a specific http status code that indicates the user the reason for the error. Show an informative alert in the client.