URL blacklist logic no longer gets confused with urls that are a substring but not actually a subdomain of a blacklisted url.
E.g. with the old logic if "hoy.es" is a blacklisted URL and a user tried to subscribe to "linuxhoy.es", a BlacklistedUrlError would be raised preventing him from subscribing. With the updated logic no error is raised. On the other hand if the user tried to subscribe to "linux.hoy.es" an error would be raised as before, which is the expected behavior.

To achieve this the naive substring mathching has been replaced with a more sophisticated regex pattern matching based on the URL host (extracting the URL host requires parsing the url, adding the http uri-scheme beforehand if necessary).