When checking a url against the blacklist, return false if the passed string is blank or nil.
This means that blank/nil urls are never considered blacklisted, which makes sense in a way.

This should help with a bug that raised an error when checking if a feed with a nil url was blacklisted (Feed#url is allowed to be nil, while Feed#fetch_url cannot be nil).