Control the situation in which a blacklisted URL is present in the OPML file submited by the user for subscriptions import.
Previously the specific BlacklistedUrlError raised was not controlled; the end result was similar (the feed was not subscribed and a failure was added to the opml import process) but it's nicer to control it and log a nicer error specific to this situation. Also no error backtrace is logged, it is just clutter in this case.