Improved log message.
The log message made reference to the user having submitted the subscribe form. This is no longer necessarily true, the method can also invoked as part of an opml import.

In general it's better that log messages make reference to what's happening in the particular class. A log message in a class that made reference to what supposedly was happening in a vew was a bad idea.