Pull ng-infinite-scroll with bower instead of copying to /vendor/assets by hand.
It pulls angularjs as a dependency, so I'm also commiting the latest version of angularjs in /vendor/assets/bower_components. However this angularjs is only there to satisfy the bower dependency, it won't be included in the asset pipeline and we will never serve it from the webserver (we're linking angularjs from a CDN instead).