Pull uri.js with bower, instead of copying to /vendor/assets by hand.
The version with "URIjs" name is installed, instead of the one named "uri.js". As far as I can tell they are the same, but the "uri.js" name does not mix well with the asset pipeline (it gives errors about no file named "uri.js" existing).

It pulls jquery as a dependency, but we won't add it to the asset pipeline and it will never be served from the webserver, it is there just to satisfy the bower dependency requirement. The app links jquery from a CDN.