Replaced in Bowerfile "within-viewport" js library with "isInViewport".
Within-viewport doesn't seem to be under active development and its latest version worked weirdly (it seems it doesn't correctly detect the viewport coordinates). On the other hand isInViewport is recent and seems to work fine.

Also the code that only fires certain events only after the user stops scrolling for 250ms has been rewritten. Now it doesn't store the timer in an attribute of the root scope, instead a local variable in the service is used. Angularjs services are singletons, so the variable is guaranteed to be persistant between successive invocations of the methods.