Store the url blacklist as a list of host objects during initialization, instead of a list of strings with URLs.
This should improve performance: every time a feed is saved it is checked against the blacklist, and at the comparison is made between host objects. Having the blacklisted host objects already in memory avoids having to parse, add uri-scheme and extract the host for each blacklisted url every time a feed is saved.