The worker that runs when an OPML import finishes is in the same queue as the rest of workers related to OPML import.
This worker sets the job state to either SUCCESS or ERROR, and sends a notification email.

The queue it is in now has a higher priority than the maintenance and update_feeds queues (but less than the interactive queue). Previously this worker was in the update_feeds queue, which meant that after the import had finished it could take some time until the job state was set to SUCCESS or ERROR and the email sent. Now it should happen immediately after the import finishes.