When subscribing to a feed, check if the feed is blacklisted before checking if the feed is already in the DB.
In the case that the feed is blacklisted AND exists in the DB (this should never happen, but assume it does because of a bug) the user should be blocked from subscribing. This is the current behavior; before this commit the user was allowed to subscribe to the existing feed.