If FeedClient.fetch is passed a false in the http_caching argument, forcibly disable Rack::Cache from the RestClient components.
I think that RestClient behaves like a singleton, and once Rack::Cache is enabled it will be enabled for other invocations of FeedClient.fetch, even if false is passed in http_caching.

To make sure that the caching behavior is as expected, Rack::Cache is:

- enabled if true is passed in the http_caching argument
- disabled (RestClient.disable) if false is passed in the http_caching argument

To minimize possible synchronization issues between sidekiq workers simultaneously invoking FeedClient.fetch, the code that enables or disables Rack::Cache has been moved immediately before the call to RestClient.get so that as little time as possible passes between them.