Fix a bug in which scheme-relative URLs in entries were treated as relative URLs.
This caused the feed host and URI-scheme to be prepended to the URI in the entry, which often resulted in invalid URLs (images that didn't load etc).

Generally speaking, scheme-relative URLs should be left untouched when processing entries. Only manipulations possibly worthy are trimming and converting internationalized URLs to punycode.