CI environment loads some CSS files (bootstrap, fontawesome) from a CDN, instead of compiling them with the asset pipeline.
This is similar to commit 6324526078d89a562bc5aa4431a25507da8e9513 but for CSS instead of JS assets.

This should speed up the asset pipeline in the CI environment, reducing the number of times acceptance tests fail because of a timeout.

Again, I'm not happy about having a CI environment which is different from production, but I'm even less happy having acceptance tests that randomly fail.

Ideally CSS would be loaded from a CDN with a fallback to a local asset. Fallback.js does not implement this properly for now, detecting if a CSS has loaded is not trivial. They will probably address it in the next version.