Serve some popular assets (jquery, angularjs, bootstrap and fontawesome) from a CDN instead of as local assets.
This means the CI environment is again the same as the production env.

The lines that load those assets as local instead of the CDN are still present, but commented. The included instructions in the comments indicate how to revert this change and quickly serve the assets locally, in case a CDN goes down.