Published attribute of EntryState model has its value fixed to the "published" attribute of the associated Entry instance.
No matter what value other code tries to set, it is guaranteed that an EntryState instance has the same published value as its associated Entry instance.

Also, when creating EntryStates (when adding a feed subscription or when adding new entries to a feed), use the create! method, so that any error are raised to force me to fix them, instead of being silenced.