When retrieving paginated entries, set SQL order conditions on entry_states table.
This should help the database choose an execution query that gets results from entry_states as soon as possible, avoiding nested loops with entries.

For the moment the entry creation datetime is no longer used to order entries with the same publishing date. The Entry id is used instead, which as long as is a sequential number should be enough.