Fixed migrations.
The migrations I'm trying to apply use the EntryState model. In its current state it makes reference to the new "published" and "entry_created_at" attributes. This means that for the deployment to run successfully, both new columns must be added during the migration before using the EntryState model to give default values to these columns.