When the user clicks on "read all" button, the unread count for affected feeds is updated as part of the request-response cycle, instead of enqueuing a job to perform this.
The old way of enqueuing a job meant a faster response, but if the user refreshed the page or clicked on the folder again outdated unread counts could appear again in the sidebar. This was surprising for users, in a bad way.

The sidekiq worker UpdateFeedUnreadCountWorker is no longer enqueued anywhere, so it's been deleted.