The first confirmation reminder email is sent only once to a user.
To make sure it is not sent more than once, a boolean flag has been added to the User model to indicate if it has already been sent for a given user. Also an index has been added to speed up the query that selects users to send the first reminder.