New possible error fixed when fixing text encoding: strings that are valid UTF-8 but have other encodings.
This can happen with feed or entry fields (title etc). If the string is valid UTF-8 but has another encoding, the UTF-8 encoding is forced.

This sometimes happens because RestClient does not detect the encoding correctly from the content-type HTTP header, and returns strings in other encodings that were supposed to be UTF-8 (and the HTTP headers correctly indicated it, but were ignored). It seems to be fixed for RestClient version 2.0.0, which is currently unreleased.

When a RestClient version that handlses encodings better is released, perhaps some of these hacks will be no longer necessary.