Added a bit of entropy to schedules, so that scheduled updates are a bit more evenly distributed over time.
The old implementation tended to concentrate updates at specific moments during the day, which is a suboptimal allocation of server resources. Now every time the adaptative scheduling algorithm runs, a random number of seconds, up to 1 minute, is added or substracted to the new update interval for the feed. This should help distributing updates over time.