Do not lose subscriptions because of autodiscovery.
If during autodiscovery it is found that another feed already exists with the autodiscovered URL, the feed being fetched is destroyed and the old one returned instead. This is intended to detect duplicate feeds and keep a single feed in the DB, even if this is only detected after autodiscovery.

If this happens when the first user subscribes to the feed being fetched, this works fine. However if this happens later during a scheduled update (because the fetch_url is no longer valid, feedbunch performs autodiscovery on the html page and finds a url that already belongs to another feed in the DB), if we simply destroy the feed being updated the subscribed users are simply deprived of the subscription.

The change here is to gather users subscribed to the feed being fetched, and subscribe all of them to the old autodiscovered feed before destroying the fetched feed. This way users should simply see entries from the already existing feed in the app.