Improved performance of ResetDemoUserWorker.
Instead of just unsubscribing from all feeds and destroying all folders just to recreate them all again, now the worker looks at the subscriptions and folders of the demo user, compares them with the configured defaults, and makes just the necessary changes. Also all entries are marked as unread for the demo user with a mass update, for performance reasons.

The price to pay is that the worker is now much more complex.