Experimental: when a scheduled update runs, skip looking for and removing other scheduled updates for the same feed, before scheduling the next update.
My profiling seems to indicate that searching enqueued scheduled updates for the same feed and removing them from Redis is extremely expensive (about half the running time for ScheduledUpdateFeedWorker). Removing it should lighten the load on the background server.

This means that jobs are no longer idempotent. In particular manually running ScheduledUpdateFeedWorker could end up scheduling multiple updates for the feed in the future, which is a waste of resources. A solution will have to be implemented if this fix proves to be good enough to keep in production.