Performance improvement when fixing schedules.
Querying Redis for each feed is very slow. What we do instead is to get all IDs passed to scheduled updates, all at once for each queue. We have then an array of IDs in a ruby variable for each named queue; it's then a matter of comparing available feed IDs with the IDs in each array. This means that there are less Redis queries, most operations are performed in Ruby.

Further improvements are possible: instead of iterating with find_each over available feeds using ActiveRecord, it would be possible to get an array of available feed IDs with a single SQL query, and then just compare array values. I don't know yet if the performance gained would be worth it.