When exporting OPML subscriptions, do not attempt to order feeds and folders by title.
The .order clauses was being ignored by the find_each anyway, removing them does not change how it was already working.

As a result the exported OPML will be less nice to read for human eyes, but it's intended to be imported by another reader software anyway and the contents do not change, only their order.