WIP: keyboard shortcuts.
Moving the small caret that indicates the currently highlighted entry is done with jquery. AngularJS ng-show is too slow, it performs for each entry and the delay feels very unresponsive to the user. The state (which entry is highlighted) is still kept with angularjs, but the hiding-showing of the carets is done with jquery. This also forces me to add a CSS rule to show the caret of the first entry when loading the first page of entries; I cannot trust jquery to show the first caret in this case because often the jquery code runs before angularjs has had time to actually display entries. This is a case in which having jquery and angularjs code each living at its own pace is a bit of a problem, but I see no better way. As stated above angularjs is just too slow in this case.