When using keyboard shortcuts to change the highlighted entry, the entries list autoscrolls to show the entry in the viewport if necessary.
Also while autoscrolling, highlighting entries by hovering the mouse over them is disabled. This is necessary because otherwise, if autoscrolling entries is triggered while the mouse is on the entries list, this would trigger a hover event and at the end of the animation the highlighted entry would be the one that happens to be under the mouse pointer. To prevent this, the handler of mouseover event is disabled temporarily while the autoscroll animation runs.