Temporarily disable highlighting entries by mouseover after using keyboard shortcuts to open or close an entry.
This is to prevent the scrolling animation triggering a mouseover event that ends up highlighting the entry that happens to be under the mouse pointer, which is not the expected behavior.

I'm not too happy this this yet. Opening an entry with the spacebar currently starts a 1-second interval during which moving the mouse does not highlight entries. This prevent a user from moving the mouse and opening entries with the spacebar at the same time, which could be surprising for the poor user.

Perhaps it would be better to reenable the "highlight on mouseover" behavior not after a timeout, but if the user moves the mouse. This would prevent mouseover events triggered by autoscrolls, while at the same time still allowing the user to immediately use the mouse to highlight entries.