When a keyboard shortcut is used, the "highlight on mouseover" feature is disabled.
This prevents accidental mouseover events when an autoscroll animation is triggered (e.g. by opening an entry).

When the user moves the mouse, the "highlight on mouseover" feature is immediately enabled again. This allows switching from keyboard to mouse without a hitch.

This replaces the old implementation, in which mouseovers were temporarily disabled for a few milliseconds. This was hard to get right, and could end up having a user moving the mouse while mouseovers were still disabled. Also this implementation is much cleaner.