WIP: keyboard shortcuts to navigate up and down the sidebar.
The highlighted state is no longer managed with :hover CSS rules. Instead an angularjs service has been implemented and the state (which link in the sidebar is highlighted) is managed with javascript and CSS classes instead.

This is a preliminary step before implementing keyboard shortcuts to move up-down the highlighted sidebar link.