When setting the currently selected feed, most operations that were performed by the angularjs controller have been moved to currentFeedSvc.set()
This means better reusability: now currentFeedSvc.get can be invoked from keyboardShortcutsSvc, to enable a keyboard shortcuts that selects the current feed for reading, and the results will be the same as if the user clicked on the feed link with the mouse. This is not yet implemented.

A similar change will be implemented when selecting the current folder.