When fetching a feed, assume response is in the charset reported in the HTTP content-type header, and convert to utf-8.
This fixes an error in which sometimes a response with encoding different from utf-8 returned true to .valid_encoding? but raised an error when actually trying to convert it to utf-8. I think this is a Ruby bug and that the response string should have returned a false to .valid_encoding? in this case.