Control the case in which user is already subscribed to feed, when autodiscovery determines that feed being fetched is a duplicate.
When fetching a feed, it autodiscovery determines that the feed being fetched is a duplicate of another already existing feed, the feed being fetched is destroyed and all subscribed users are subscribed to the already existing (original) feed.

The change is: before subscribing each user to the already existing feed, check if he's already subscribed to it. Attempting to subscribe a user to an already subscribed feed raises an error and causes the process to fail.