Move the check to see if a particular tour should be shown to the tourSvc service.
This way it is more self-contained. Code that can potentially trigger showing a tour just invokes the service, and it is the responsability of the service to check user config flags to see if the tour has to be actually shown. Code outside the service does not need to look at the value of these flags.