When the user toggles showing/hiding read entries, the start link in the sidebar becomes the highlighted sidebar link.
This is desirable because hiding read entries can mean that feeds disappear from the sidebar; in particular the currently highlighted feed can disappear from the sidebar, which causes undefined behavior. To avoid this, each time this is toggled by the user the start link (which is guaranteed to be in the sidebar) is highlighted.