Fixes for the case in which the correct encoding is reported in HTTP content-type header, but the wrong encoding is reported in the XML itself.
I've found this happens with some chinese feeds: they are encoded in GBK, and the HTTP content-type header correctly indicates it, but the feed XML itself says that the encoding is utf-8 (which it isn't).

In this case Feedjira parses the feed but in all output strings (feed title, entry content, etc) the encoding is set to the one reported in the XML; in the above case, the encoding is set to utf-8 even though the internal representation is GBK for those strings. This causes problems later when attempting to save in the database, either an error is raised or useless gibberish is saved in the db.

The fix is to pass the encoding detected when fetching the feed to all methods that deal with the parsed feed and its entries. Each time one of those strings that could potentially be affected by this problem is handled, it is checked to see if it's valid for its current encoding; if it isn't, the encoding detected when fetching the feed is forced instead (and hopefully it is valid now). This way the correct encoding should be reported (and activerecord callbacks will take care of converting all strings to utf-8 prior to actually saving them in the db, for consistency).

It is a bit hacky, cumbersome and annoying. Definitely this situation is a bug in the feed, but if I want users to be able to subscribe transparently to feeds with this kind of bugs I'm afraid all those encoding checks and conversions have to be in place in Feedbunch.