Control an error when fetching a feed: the server reports zipped content in the HTTP headers, but the response is not actually zipped.
In this case a Zlib::GzipFile::Error is raised. This error was not controlled, so the error was raised to Sidekiq and the worker was put in the Retries queue.

Now the error is controlled, the feed is marked as failing (which ends up making the feed unavailable if it keeps failing for too long) and the worker is successful as far as Sidekiq is concerned.

Also some error classes that were not controlled when manually subscribing or refreshing a feed have been added to their rescue clauses. Those errors were already being rescued in scheduled feed updates.